#240 文字の影の幅が以前より狭くなっていた問題を修正
